Combined layouts 3 and 4.
