+++ scholly/base/src/layouts/layout3/components/achievementsPage/Achievements.js
@@ -22,12 +22,12 @@ const Achievements = () => {
         </title>
       </Helmet>
       <Container className="pages-top-padding text-center mt-3 mb-3">
-        <div className="team-pg-title">Achievements</div>
+        <div className="team-pg-title w-100 text-center">Achievements</div>
       </Container>
       <Container fluid className="team-card-container">
         <CardDeck className="team-card-deck">
-          {teamAchievements.map((achievement) => (
-            <Achievement achievement={achievement} key={achievement._id} />
+          {teamAchievements.map((achievement, i) => (
+            <Achievement style={{ display: 'flex', flexDirection: i % 2 === 0 ? 'row' : 'row-reverse' }} achievement={achievement} key={achievement._id} />
           ))}
         </CardDeck>
       </Container>
