--- scholly/base/src/layouts/layout4/App.js
@@ -1,36 +1,0 @@
-/**
- * Root App.js
- */
-import React from 'react';
-import { Switch, Route } from 'react-router-dom';
-import { Helmet } from 'react-helmet';
-import Header from './components/layout/Header';
-import getRoutes from './components/router/routes';
-import './components/centered.css';
-import { TEAM_INFO } from '../../global/data';
-
-const App = () => {
-  const { teamName } = TEAM_INFO;
-  const routeItems = getRoutes().map(({ path, exact, component }) => {
-    const View = component;
-    return (
-      <Route exact={exact} path={path} key={path}>
-        <div>{View ? <View /> : null}</div>
-      </Route>
-    );
-  });
-
-  return (
-    <>
-      <Helmet>
-        <title>{teamName}</title>
-      </Helmet>
-      <Header />
-      <Switch>
-        {routeItems}
-      </Switch>
-    </>
-  );
-};
-
-export default App;
