--- scholly/base/src/layouts/layout4/components/twitter/TwitterFeed.css
@@ -1,5 +1,0 @@
-.twitter-feed {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
