--- scholly/base/src/layouts/layout4/components/twitter/TwitterLink.css
@@ -1,20 +1,0 @@
-.twitter-link {
-    position: fixed;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    align-items: stretch;
-}
-
-.twitter-link > * {
-    height: 100%;
-}
-
-.twitter-link_link_message {
-    margin-bottom: 10px;
-}
-
-.twitter-link_button {
-    margin-top: 10px;
-    margin-bottom: 10px;
-}
